country-exchange-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
aiomysql
pydantic>=2.0
python-dotenv
httpx
pillow
matplotlib